consolidating client script files
