--- src/client/public/index.js
@@ -1,7 +1,0 @@
-document.addEventListener('DOMContentLoaded', () => {
-  const button = document.querySelector('button');
-  button.addEventListener('click', () => {
-    // Your code here, e.g., redirecting to the login page
-    window.location.href = './login.html';
-  });
-});
