--- src/client/public/js/formEvents.js
@@ -1,21 +1,0 @@
-// document.querySelector('#p1').addEventListener('change', function() {
-//     document.querySelector('#p2').pattern = this.value;
-// });
-
-// function checkpass(e) {
-//     e.preventDefault(); // Use the event parameter
-//     const form = document.querySelector('form[action="signUp.js"]');
-//     const p1 = document.querySelector('#p1').value;
-//     const p2 = document.querySelector('#p2').value;
-
-//     if (p1 !== p2) {
-//         alert('Passwords do not match!');
-//         return false;
-//     }
-//     return true;
-// }
-
-// Add the submit event listener outside the function
-// document.querySelector('form[action="signUp.js"]').addEventListener('submit', function(e) {
-//     checkpass(e);
-// });
