+++ src/client/public/js/signUp.js
@@ -1,46 +1,102 @@
-const form = document.getElementById('register_submit');
 
-async function handleClick(event) {
-  event.preventDefault();
+function initPasswordReset() {
+  const form = document.querySelector('form[action="resetPassword.js"]');
 
-  const name = document.getElementById('name').value;
-  const email = document.getElementById('email').value;
-  const password = document.getElementById('password').value;
-  const reenterPassword = document.getElementById('reenter-password').value;
-
-  // Perform validation checks once
-  if (!name || !email || !password || !reenterPassword) {
-    alert('Please fill in all fields.');
+  if (!form) {
     return;
   }
 
-  if (password !== reenterPassword) {
-    alert('Passwords do not match.');
-    return;
+  function handlePasswordReset(e) {
+    e.preventDefault();
+
+    const formData = {
+      email: form.email.value,
+      oldPassword: form.pass_old.value,
+      newPassword: form.pass_new.value
+    };
+
+    fetch('/reset-password', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(formData)
+    })
+    .then(function(response) {
+      if (response.ok) {
+        return response.json().then(function(data) {
+          console.log('Password reset successful:', data.message);
+          // You could display the success message to the user
+          // For example:
+          alert(data.message || 'Password successfully reset');
+          // window.location.href = '/login';
+        });
+      } else {
+        return response.json().then(function(data) {
+          throw new Error(data.message || 'Failed to reset password');
+        });
+      }
+    })
+    .catch(function(error) {
+      console.error('Password reset failed:', error.message);
+      // Display error to user
+      alert(error.message);
+    });
   }
 
-  // Create data object and send request
-  const data = {
-    username: name,
-    email: email,
-    password: password
-  };
-
-  const response = await fetch('/api/v1/register', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify(data)
-  });
-
-  if (response.status === 201) {
-    // Redirect the user to a success page
-    window.location.href = '/dashboard';
-  } else {
-    const data = await response.json();
-    alert(data.message);
-  }
+  form.addEventListener('submit', handlePasswordReset);
 }
 
-form.addEventListener('click', handleClick);
+function initSignUp() {
+  const submitButton = document.getElementById('register_submit');
+  if (!submitButton) {
+    return;
+  }
+
+  async function handleClick(event) {
+    event.preventDefault();
+
+    const name = document.getElementById('name').value;
+    const email = document.getElementById('email').value;
+    const password = document.getElementById('password').value;
+    const reenterPassword = document.getElementById('reenter-password').value;
+
+    // Perform validation checks once
+    if (!name || !email || !password || !reenterPassword) {
+      alert('Please fill in all fields.');
+      return;
+    }
+
+    if (password !== reenterPassword) {
+      alert('Passwords do not match.');
+      return;
+    }
+
+    // Create data object and send request
+    const data = {
+      username: name,
+      email: email,
+      password: password
+    };
+
+    const response = await fetch('/api/v1/register', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(data)
+    });
+
+    if (response.status === 201) {
+      // Redirect the user to a success page
+      window.location.href = '/dashboard';
+    } else {
+      const data = await response.json();
+      alert(data.message);
+    }
+  }
+
+  submitButton.addEventListener('click', handleClick);
+}
+
+export { initPasswordReset, initSignUp };
