+++ src/client/public/register.html
@@ -18,7 +18,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet">
     <link rel="manifest" href="manifest.json" crossorigin="use-credentials">
-    <link rel="stylesheet" type="text/css" href="../public/css/register.css">
+    <link rel="stylesheet" type="text/css" href="../css/register.css">
   </head>
 
   <body>
@@ -54,10 +54,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.2/js/all.min.js"
       integrity="sha512-8pHNiqTlsrRjVD4A/3va++W1sMbUHwWxxRPWNyVlql3T+Hgfd81Qc6FC5WMXDC+tSauxxzp1tgiAvSKFu1qIlA=="
       crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-    <script src="/js/signUp.js"></script>
-    <script src="/js/formEvents.js"></script>
-    <script src="/js/resetPassword.js"></script>
-    <script src="/app.js"></script>
+    <script type="module" src="./app.js"></script>
   </body>
 
 </html>
